feat Apache POI Excel画图表

+++ weibo-service-service/src/main/java/top/deepdesigner/weibo/weiboservicebackend/service/utils/RedisUtil.java
@@ -50,6 +50,6 @@ public class RedisUtil {
     }
 
     public static long randomExpire(int min, int max) {
-        return min + RANDOM.nextLong() * (max - min);
+        return (long) (min + RANDOM.nextFloat() * (max - min));
     }
 }
